pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-the_local_minima"
version = "0.1.0"
description = "The Local Minima environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn",
    "pydantic",
    "numpy",
    "pvlib",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m the_local_minima.server.app
server = "the_local_minima.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["the_local_minima", "the_local_minima.server"]
package-dir = { "the_local_minima" = ".", "the_local_minima.server" = "server" }